requirements_local.txt
# Streamlit 프론트엔드
streamlit==1.28.1

# HTTP 요청
requests==2.31.0

# 오디오 처리
streamlit-audiorecorder==0.0.5

# LangChain 관련 (로컬 테스트용)
langchain
langchain-community
faiss-cpu

# 추가 유틸리티
python-dotenv==1.0.0

# 데이터 처리
pandas==2.1.4
numpy==1.24.3

# 이미지 처리 (필요시)
Pillow==10.1.0

# JSON 처리
orjson==3.9.10 